fixup: propagate rename narrator -> flowInspector
