+++ packages/async-flow/test/async-flow-no-this.test.js
@@ -29,7 +29,7 @@ const testPlay = async (t, zone) => {
   const { guestFunc } = {
     async guestFunc() {
       // @ts-expect-error TS does not yet understand guest contexts
-      const { narrator, self, state, facets } = this;
+      const { flowInspector, self, state, facets } = this;
       t.is(self, undefined);
       t.is(state, undefined);
       t.is(facets, undefined);
